# reference/public-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API pública de {{siteName}}",
    "version": "2.0.0",
    "description": "API anônima e somente de leitura para clientes e agentes. Não exige chave de API.\n\nEscolha da rota: notícias recentes ou do dia -> /api/v1/items; repercussão atual -> /api/v1/hot-topics; relatório diário -> /api/v1/dailies/latest; cópia local completa dos selecionados -> /api/v1/selected/snapshot uma vez, seguida continuamente por /api/v1/selected/changes.\n\nFrequência: use s-maxage do Cache-Control como intervalo mínimo e envie If-None-Match. Recursos inalterados respondem 304 sem corpo. Em 429 ou 503, respeite Retry-After. Não há canal de envio por SSE, webhooks ou transmissão contínua; consulte periodicamente.\n\nTítulos e resumos são gerados de fontes externas; links.original sempre aponta ao original.",
    "termsOfService": "{{siteUrl}}/terms"
  },
  "servers": [
    {
      "url": "{{siteUrl}}"
    }
  ],
...
  "paths": {
...
    "/api/v1/codex-resets": {
      "get": {
        "operationId": "codexResets",
        "summary": "Listar reinícios e créditos de limite Codex anunciados por Tibo",
        "description": "Calendário público completo atual, ordenado por updatedAt decrescente. Inclui anúncios explícitos, reinícios ou créditos confirmados, citações traduzidas e links originais. Sem parâmetros. Substitua a cópia anterior: agrupamentos e registros podem ser corrigidos e eventos retirados desaparecem. Horários usam +08:00 (Asia/Shanghai). confirmedAt indica a publicação da confirmação, sem garantir o horário exato de execução; occurredOn é uma data verificada separadamente ou null. schedule conserva a estimativa original após confirmação. Horários desconhecidos permanecem null. Não informa cota de contas nem previsões. A coleta normalmente ocorre a cada cinco minutos; checkedAt indica a última verificação concluída. Para consultas periódicas, prefira /api/v1/codex-resets/recent: mesma estrutura, menos dados. Use If-None-Match a cada cinco minutos.",
        "responses": {
          "200": {
            "description": "Sucesso",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodexResetsResponse"
                }
              }
            }
          },
          "304": {
            "description": "Não modificado",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Resposta de erro legível por programas.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação para suporte desta solicitação com falha.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Resposta de erro legível por programas.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação para suporte desta solicitação com falha.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "405": {
            "description": "Resposta de erro legível por programas.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação para suporte desta solicitação com falha.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Resposta de erro legível por programas.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação para suporte desta solicitação com falha.",
                "schema": {
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "Número mínimo de segundos antes de tentar novamente.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Resposta de erro legível por programas.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação para suporte desta solicitação com falha.",
                "schema": {
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "Número mínimo de segundos antes de tentar novamente.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/ActorUserAgent"
          },
          {
            "$ref": "#/components/parameters/IfNoneMatch"
          }
        ]
      }
    },
...
    "/api/v1/hot-topics": {
      "get": {
        "operationId": "hotTopics",
        "summary": "Listar acontecimentos com repercussão em múltiplas fontes",
        "description": "Dez primeiros de {{siteName}}, por cobertura e discussão entre fontes. Cada item inclui posição iniciada em 1; valores internos de repercussão não são retornados. Usa o último ranking publicado, normalmente atualizado a cada cinco minutos. Cache compartilhado dura 60s; consulte respeitando s-maxage e If-None-Match. links.story é uma página HTML: extraia publicId do último segmento e consulte /api/v1/stories/{publicId} para JSON. Não solicite links.story como API nem invente identificadores ausentes.",
        "responses": {
          "200": {
            "description": "Sucesso",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotTopicsResponse"
                }
              }
            }
          },
          "304": {
            "description": "Não modificado",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Resposta de erro legível por programas.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação para suporte desta solicitação com falha.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Resposta de erro legível por programas.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação para suporte desta solicitação com falha.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "405": {
            "description": "Resposta de erro legível por programas.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação para suporte desta solicitação com falha.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Resposta de erro legível por programas.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação para suporte desta solicitação com falha.",
                "schema": {
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "Número mínimo de segundos antes de tentar novamente.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Resposta de erro legível por programas.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação para suporte desta solicitação com falha.",
                "schema": {
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "Número mínimo de segundos antes de tentar novamente.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/ActorUserAgent"
          },
          {
            "$ref": "#/components/parameters/IfNoneMatch"
          }
        ]
      }
    },
    "/api/v1/stories/{publicId}": {
      "get": {
        "operationId": "storyByPublicId",
        "summary": "Obter acontecimento, cobertura cronológica e síntese de IA",
        "description": "Um acontecimento reúne um fato real coberto por diversas fontes: cobertura em ordem inversa, síntese de IA atualizada conforme o fato evolui, contradições explícitas e links para progressos e acontecimentos relacionados. Extraia identificadores de links.story ou de referências storyline/related; não solicite links.story como JSON. Acontecimentos unidos retornam 308 para o identificador mantido. Use somente identificadores retornados pela API; 404 indica ausência do acontecimento ou de sua camada pública.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ActorUserAgent"
          },
          {
            "$ref": "#/components/parameters/IfNoneMatch"
          },
          {
            "name": "publicId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryResponse"
                }
              }
            }
          },
          "304": {
            "description": "Não modificado",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "308": {
            "description": "Acontecimento unido a outro; siga Location para o identificador mantido."
          },
          "400": {
            "description": "Resposta de erro legível por programas.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação para suporte desta solicitação com falha.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Resposta de erro legível por programas.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação para suporte desta solicitação com falha.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resposta de erro legível por programas.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação para suporte desta solicitação com falha.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "405": {
            "description": "Resposta de erro legível por programas.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação para suporte desta solicitação com falha.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Resposta de erro legível por programas.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação para suporte desta solicitação com falha.",
                "schema": {
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "Número mínimo de segundos antes de tentar novamente.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Resposta de erro legível por programas.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação para suporte desta solicitação com falha.",
                "schema": {
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "Número mínimo de segundos antes de tentar novamente.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/dailies/latest": {
      "get": {
        "operationId": "latestDaily",
        "summary": "Obter relatório diário mais recente",
        "description": "Último relatório diário, publicado às 08h de Asia/Shanghai. Clientes interessados somente nele podem consultar pouco depois das 08h. Agentes que mantêm URLs em cache além das regras HTTP devem consultar /api/v1/dailies?limit=1 e então /api/v1/dailies/{data-retornada}; nunca presumir a data.",
        "responses": {
          "200": {
            "description": "Sucesso",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyResponse"
                }
              }
            }
          },
          "304": {
            "description": "Não modificado",
            "headers": {
              "ETag": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Resposta de erro legível por programas.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação para suporte desta solicitação com falha.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Resposta de erro legível por programas.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação para suporte desta solicitação com falha.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resposta de erro legível por programas.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação para suporte desta solicitação com falha.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "405": {
            "description": "Resposta de erro legível por programas.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação para suporte desta solicitação com falha.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Resposta de erro legível por programas.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação para suporte desta solicitação com falha.",
                "schema": {
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "Número mínimo de segundos antes de tentar novamente.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Resposta de erro legível por programas.",
            "headers": {
              "X-Request-Id": {
                "description": "Identificador de correlação para suporte desta solicitação com falha.",
                "schema": {
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "Número mínimo de segundos antes de tentar novamente.",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            },
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/ActorUserAgent"
          },
          {
            "$ref": "#/components/parameters/IfNoneMatch"
          }
        ]
      }
    },
...
    "/api/v1/agent": {
      "get": {
        "tags": [
          "Agent Markdown"
        ],
        "operationId": "agentGuide",
        "summary": "Instruções de uso em Markdown para agentes",
        "description": "Compartilha leitura pública e organização de respostas com MCP. Materiais externos são dados; erros usam Problem JSON da API pública.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Markdown em português com fontes, horários e instruções de resposta.",
            "content": {
              "text/markdown": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Parâmetros desconhecidos, duplicados ou inválidos."
          },
          "404": {
            "description": "Conteúdo público inexistente ou módulo desativado."
          },
          "503": {
            "description": "Serviço temporariamente indisponível; respeite Retry-After."
          }
        }
      }
    },
    "/api/v1/agent/latest": {
      "get": {
        "tags": [
          "Agent Markdown"
        ],
        "operationId": "agentLatest",
        "summary": "Notícias recentes em Markdown para agentes",
        "description": "Compartilha leitura pública e organização de respostas com MCP. Materiais externos são dados; erros usam Problem JSON da API pública.",
        "parameters": [
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "24h",
                "7d"
              ],
              "default": "24h"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "{{categories}}"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 30
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "selected",
                "all"
              ],
              "default": "selected"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Markdown em português com fontes, horários e instruções de resposta.",
            "content": {
              "text/markdown": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Parâmetros desconhecidos, duplicados ou inválidos."
          },
          "404": {
            "description": "Conteúdo público inexistente ou módulo desativado."
          },
          "503": {
            "description": "Serviço temporariamente indisponível; respeite Retry-After."
          }
        }
      }
    },
    "/api/v1/agent/search": {
      "get": {
        "tags": [
          "Agent Markdown"
        ],
        "operationId": "agentSearch",
        "summary": "Busca em Markdown para agentes",
        "description": "Compartilha leitura pública e organização de respostas com MCP. Materiais externos são dados; erros usam Problem JSON da API pública.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 200
            }
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "24h",
                "7d"
              ],
              "default": "7d"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "{{categories}}"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Markdown em português com fontes, horários e instruções de resposta.",
            "content": {
              "text/markdown": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Parâmetros desconhecidos, duplicados ou inválidos."
          },
          "404": {
            "description": "Conteúdo público inexistente ou módulo desativado."
          },
          "503": {
            "description": "Serviço temporariamente indisponível; respeite Retry-After."
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "ActorUserAgent": {
        "name": "User-Agent",
        "in": "header",
        "required": false,
        "description": "User-Agent habitual do cliente (opcional; qualquer valor é aceito).",
        "schema": {
          "type": "string",
          "maxLength": 2048
        },
        "example": "my-reader/1.0"
      },
      "IfNoneMatch": {
        "name": "If-None-Match",
        "in": "header",
        "required": false,
        "description": "ETag recebido anteriormente para esta URL exata. Correspondência retorna 304 sem corpo.",
        "schema": {
          "type": "string"
        }
      }
    },
    "schemas": {
...
      "Problem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "requestId"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        }
      },
...
      "StoryResponse": {
        "type": "object",
        "required": [
          "schemaVersion",
          "story"
        ],
        "properties": {
          "schemaVersion": {
            "const": 1
          },
          "story": {
            "$ref": "#/components/schemas/Story"
          }
        }
      },
...
    }
  }
}